+++ changelog.md
@@ -5,6 +5,7 @@ v1.2 (2026-xx-xx):
 * Initialize properly if loaded mid-level
 * Improve coop HUD issues, sorting by distance w/ default 4p max
 * Add cd_hudsorttime for HUD sort interval
+* Add pinplayer2 / unpinplayer2 for splitscreen use
 * Code cleanup!
 
 v1.1 (2021-06-11):